# docs/api-document.yaml
openapi: 3.0.2
info:
  title: API定義
  description: |
    情報実験のアプリケーションサーバ用です
  version: 1.1.0
  contact:
    url: 'https://github.com/THEToilet'
servers:
  - url: 'http://localhost:8081'
tags:
  - name: chat
    description: チャット関連APIS
  - name: room
    description: 部屋関連API
paths:
  '/chat/{roomId}/message/post':
    post:
      tags:
        - chat
      summary: メッセージ送信API
      description: チャットにてメッセージを送る際に使います
      parameters:
        - name: session
          in: header
          description: 認証トークン
          schema:
            type: string
          required: true
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageSource:
                  type: string
                  description: 送信する画像ソース(base64変換されている)
                fileName:
                  type: string
                  description: ファイル名
                fileExtension:
                  type: string
                  description: ファイル拡張子情報
              required:
                - imageSource
                - fileName
                - fileExtension
        description: メッセージ
      operationId: '1'
    parameters:
      - schema:
          type: string
        name: roomId
        in: path
        required: true
        description: 部屋Id
  '/chat/{roomId}/message/update':
    get:
      tags:
        - chat
      summary: 参加してる部屋のメッセージ一覧取得API
      description: メッセージ一覧を取得します
      parameters:
        - name: session
          in: header
          description: 認証トークン
          schema:
            type: string
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: object
                    required:
                      - messageId
                      - postedUserName
                      - imageUrl
                      - stampList
                    properties:
                      messageId:
                        type: string
                        description: メッセージのId
                      postedUserName:
                        type: string
                        description: 送信したユーザ名
                      imageUrl:
                        type: string
                        description: 画像URL
                      stampList:
                        type: object
                        description: 押されているスタンプ情報集
                        required:
                          - stamp
                        properties:
                          stamp:
                            type: object
                            description: それぞれのスタンプ情報
                            required:
                              - pushedUserName
                              - stampId
                            properties:
                              pushedUserName:
                                type: string
                                description: スタンプ押したユーザの名前
                              stampId:
                                type: string
                                description: スタンプId
                required:
                  - message
        '400':
          description: Bad Request
      operationId: '2'
    parameters:
      - schema:
          type: string
        name: roomId
        in: path
        required: true
        description: 部屋Id
  '/chat/{roomId}/exit':
    get:
      tags:
        - chat
      summary: 部屋を退出する際のAPI
      description: 部屋を退出するときに使います
      parameters:
        - name: session
          in: header
          description: 認証トークン
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
      operationId: '3'
    parameters:
      - schema:
          type: string
        name: roomId
        in: path
        required: true
        description: 部屋Id
  '/chat/{roomId}/{messageId}/stamp/post':
    post:
      tags:
        - chat
      summary: スタンプ情報送信API
      description: 画像に対してスタンプを押します。
      parameters:
        - name: session
          in: header
          description: 認証トークン
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userName:
                  type: string
                  description: 押したユーザの名前（自分の名前）
                stampId:
                  type: string
                  description: 押したスタンプのId
              required:
                - userName
                - stampId
      operationId: '4'
    parameters:
      - schema:
          type: string
        name: roomId
        in: path
        required: true
        description: 部屋Id
      - schema:
          type: string
        name: messageId
        in: path
        required: true
        description: メッセージId
  '/room/{roomId}/start':
    post:
      tags:
        - room
      summary: 部屋の人数が揃ったときにロビーサーバから情報を取得するAPI
      description: コレクションアイテム一覧情報。
      parameters: []
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                users:
                  type: array
                  maxItems: 4
                  minItems: 4
                  items:
                    type: object
                    description: ユーザ情報
                    properties:
                      userId:
                        type: string
                        description: ユーザId
                      userName:
                        type: string
                        description: ユーザの名前
                      session:
                        type: string
                        description: ユーザのセッション
                categoryId:
                  type: string
                  description: カテゴリーId
      operationId: '5'
    parameters:
      - schema:
          type: string
        name: roomId
        in: path
        required: true
        description: 部屋のId
components:
  schemas: {}
  securitySchemes: {}